Securise controller Activity fro Update & Delete

+++ Application/Services/ActivityService.cs
@@ -1,4 +1,5 @@
 using MyApi.Application.Services.Interfaces;
+using MyApi.Constants;
 using MyApi.Domain.Entities;
 using MyApi.Domain.Interfaces;
 using System.Security.Claims;
@@ -50,15 +51,63 @@ namespace MyApi.Application.Services
             await _activityRepository.AddActivityAsync(activity);
         }
 
-        public async Task<bool> UpdateActivityAsync(ActivityEntity activity)
+         public async Task<bool> UpdateActivityAsync(ActivityEntity activity, int organizerId, string role)
         {
             // Logique métier, comme vérifier si l'activité est modifiable
+            ActivityEntity? activityTemp = await _activityRepository.GetActivityByIdAsync(activity.Id);
+      
+            if (activity == null)
+            {
+                throw new ArgumentException($"Activity with ID {activity.Id} not found.");
+            }
+
+            UserEntity? user = activity.Organizers.FirstOrDefault(u => u.Id == organizerId);
+
+            if (user == null && role != Roles.ADMIN)
+            {
+                throw new ArgumentException($"Role {role} is not authorized to update this activity.");
+            }
+
             return await _activityRepository.UpdateActivityAsync(activity);
         }
 
-        public async Task<bool> DeleteActivityAsync(int id)
+        public async Task<bool> ArchiveActivityAsync(int id, string role)
+        {
+            // Logique métier, comme vérifier si l'activité est modifiable
+            ActivityEntity? activity = await _activityRepository.GetActivityByIdAsync(id);
+      
+            if (activity == null)
+            {
+                throw new ArgumentException($"Activity with ID {id} not found.");
+            }
+
+            activity.Active = false;
+
+            if (role != Roles.ADMIN)
+            {
+                throw new ArgumentException($"Role {role} is not authorized to update this activity.");
+            }
+
+            return await _activityRepository.UpdateActivityAsync(activity);
+        }
+
+        public async Task<bool> DeleteActivityAsync(int id, int organizerId, string role)
         {
             // Logique métier pour valider la suppression
+            ActivityEntity? activity = await _activityRepository.GetActivityByIdAsync(id);
+
+            if (activity == null)
+            {
+                throw new ArgumentException($"Activity with ID {id} not found.");
+            }
+
+            UserEntity? user = activity.Organizers.FirstOrDefault(u => u.Id == organizerId);
+
+            if (user == null && role != Roles.ADMIN)
+            {
+                throw new ArgumentException($"Organizer with ID {organizerId} not found in activity.");
+            }
+
             return await _activityRepository.DeleteActivityAsync(id);
         }
     }
